pyproject.toml
[project]
name = "bartrify-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Huniity",email = "adrien.dejonc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "drf-extra-fields (>=3.7.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "channels (>=4.2.2,<5.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "daphne (>=4.2.0,<5.0.0)",
]

[tool.poetry]
packages = [{include = "bartrify"}]

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.10.0"
watchfiles = "^1.0.5"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bartrify.settings"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
